Added (almost) perfect functionality to confirmation page, just can't access it

+++ Classes/Card.cs
@@ -7,12 +7,12 @@
 //  Stores user credit/debit card information
 public class Card
 {
-    public int Number {  get; set; }
+    public long Number {  get; set; }
     public int Month { get; set; }
     public int Year { get; set; }
     public string Name { get; set; }
     public int CVV { get; set; }
-    public Card(int number, int month, int year, string name, int cVV)
+    public Card(long number, int month, int year, string name, int cVV)
     {
         Number = number;
         Month = month;
